Refactored SettingsFrom.spec.js with all PR comments Implemented
