+++ tests/unit/components/playback/controls/SettingsFormCustomerData.js
@@ -1,0 +1,59 @@
+export const array_of_userid_1 = [
+  {
+    user_id: 0,
+    uuid: "2VSckkBYr2An3dqHEyfRRE",
+    nickname: "User account -1",
+  },
+  {
+    user_id: 1,
+    uuid: "5FY8KwTsQaUJ2KzHJGetfE",
+    nickname: "User account -2",
+  },
+];
+
+export const array_of_userid_2 = [
+  {
+    user_id: 0,
+    uuid: "2VSckkkkk2An3dqHEyfRRE",
+    nickname: "Lab User  -1",
+  },
+  {
+    user_id: 1,
+    uuid: "5FY8ghtsQaUJ2KzHJGetfE",
+    nickname: "Intern -1",
+  },
+];
+
+export const array_of_customerids = [
+  {
+    cust_id: 0,
+    uuid: "4vqyd62oARXqj9nRUNhtLQ",
+    api_key: "941532a0-6be1-443a-a9d5-d57bdf180a52",
+    nickname: "Customer account -1",
+    user_ids: array_of_userid_1,
+  },
+  {
+    cust_id: 1,
+    uuid: "6cBaidlJ84Ggc5JA7IYCgv",
+    api_key: "941532a0-6be1-443a-cdee-d57bdf180a52",
+    nickname: "Customer account -2",
+    user_ids: array_of_userid_2,
+  },
+];
+
+export const array_of_customerids_one = [
+  {
+    cust_id: 0,
+    uuid: "4vqyd62oARXqj9nRUNhtLQ",
+    api_key: "941532a0-6be1-443a-a9d5-d57bdf180a52",
+    nickname: "Customer account -1",
+    user_ids: array_of_userid_1,
+  },
+  {
+    cust_id: 1,
+    uuid: "6cBaidlJ84Ggc5JA7IYCgv",
+    api_key: "941532a0-6be1-443a-cdee-d57bdf180a52",
+    nickname: "Customer account -2",
+    user_ids: [],
+  },
+];
